backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
httpx==0.24.1
pydantic==2.3.0
python-multipart==0.0.6
groq==0.4.0
google-cloud-texttospeech==2.14.1
google-cloud-speech==2.21.0
pydub==0.25.1
ffmpeg-python==0.2.0